+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `snakelizeKeys` function
  - `hash`, `compare` bcrypt functions
  - `randomAfanumeric` function
+ - `transacted` function
 
 ### Changed
  - use `ts-jest` instead of `jest`